+++ src/GRA.Controllers/ViewModel/MissionControl/Participants/HouseholdListViewModel.cs
@@ -19,7 +19,7 @@ namespace GRA.Controllers.ViewModel.MissionControl.Participants
         public string UserSelection { get; set; }
         public int ActivityAmount { get; set; }
         public string ActivityMessage { get; set; }
-        public bool ShowSecretCode { get; set; }
+        public bool DisableSecretCode { get; set; }
         public string SecretCode { get; set; }
         public string SecretCodeMessage { get; set; }
         public bool ShowVendorCodes { get; set; }