+++ src/GRA.Data/Context.cs
@@ -168,6 +168,7 @@ namespace GRA.Data
         public DbSet<Model.SchoolDistrict> SchoolDistricts { get; set; }
         public DbSet<Model.SchoolType> SchoolTypes { get; set; }
         public DbSet<Model.Site> Sites { get; set; }
+        public DbSet<Model.SiteSetting> SiteSettings { get; set; }
         public DbSet<Model.System> Systems { get; set; }
         public DbSet<Model.Trigger> Triggers { get; set; }
         public DbSet<Model.TriggerBadge> TriggerBadges { get; set; }