+++ src/GRA.Domain.Model/SiteSettingFormat.cs
@@ -1,0 +1,7 @@
+﻿namespace GRA.Domain.Model
+{
+    public enum SiteSettingFormat
+    {
+        Boolean
+    }
+}