+++ src/GRA.Domain.Repository/ISiteSettingRepository.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+using GRA.Domain.Model;
+
+namespace GRA.Domain.Repository
+{
+    public interface ISiteSettingRepository : IRepository<SiteSetting>
+    {
+        Task<ICollection<SiteSetting>> GetBySiteIdAsync(int siteId);
+    }
+}